+++ src/components/offset/offset.ts
@@ -3,7 +3,7 @@ import { CuiUtils } from "../../core/models/utils";
 import { CuiHandler } from "../../app/handlers/base";
 import { CuiScrollListener, CuiScrollEvent } from "../../core/listeners/scroll";
 import { ICuiComponentAction, CuiActionsListFactory } from "../../core/utils/actions";
-import { getIntOrDefault, isStringTrue } from "../../core/utils/functions";
+import { are, getIntOrDefault, isStringTrue } from "../../core/utils/functions";
 import { EVENTS } from "../../core/index";
 
 /**
@@ -18,6 +18,10 @@ import { EVENTS } from "../../core/index";
 
 export interface CuiOffsetEvent {
     matches: boolean;
+    offsetX: number;
+    offsetY: number;
+    limitX: boolean;
+    limitY: boolean;
     timestamp: number;
 }
 
@@ -31,7 +35,7 @@ export interface CuiOffsetAttribute {
 
 export class CuiOffsetArgs {
     target: string;
-    action: ICuiComponentAction[];
+    action: string;
     offsetY?: number;
     offsetX?: number
     root: boolean;
@@ -40,11 +44,12 @@ export class CuiOffsetArgs {
         this.offsetY = 0;
         this.target = null;
         this.root = false;
+        this.action = null;
     }
 
     parse(args: any) {
         this.target = args.target;
-        this.action = CuiActionsListFactory.get(args.action);
+        this.action = args.action;
         this.offsetX = getIntOrDefault(args.offsetX, -1);
         this.offsetY = getIntOrDefault(args.offsetY, -1);
         this.root = isStringTrue(args.root)
@@ -71,6 +76,11 @@ export class CuiOffsetHandler extends CuiHandler<CuiOffsetArgs> {
     #target: Element;
     #utils: CuiUtils;
     #matched: boolean;
+    #action: ICuiComponentAction[];
+    #prevX: number;
+    #prevY: number;
+    #threshold: number;
+    #root: Element;
 
     constructor(element: Element, utils: CuiUtils, attribute: string) {
         super("CuiOffsetHandler", element, attribute, new CuiOffsetArgs(), utils);
@@ -79,6 +89,11 @@ export class CuiOffsetHandler extends CuiHandler<CuiOffsetArgs> {
         this.#target = this.element;
         this.#utils = utils;
         this.#matched = false;
+        this.#action = null;
+        this.#prevX = 0;
+        this.#prevY = 0;
+        this.#threshold = 20;
+        this.#root = null;
     }
 
     onInit(): void {
@@ -99,7 +114,9 @@ export class CuiOffsetHandler extends CuiHandler<CuiOffsetArgs> {
     }
 
     private parseAttribute() {
+        this.#root = this.getRoot();
         this.#target = this.args.target ? this.getRoot().querySelector(this.args.target) : this.element;
+        this.#action = CuiActionsListFactory.get(this.args.action);
         this.checkAndPerformActions(this.element.scrollTop, this.element.scrollLeft);
 
     }
@@ -111,25 +128,65 @@ export class CuiOffsetHandler extends CuiHandler<CuiOffsetArgs> {
 
     private checkAndPerformActions(top: number, left: number) {
         let matchesOffset = this.matchesOffset(top, left);
-        if (matchesOffset && !this.#matched) {
-            this.args.action.forEach(action => action.add(this.#target, this.#utils));
-            this.#matched = true;
-            this.callEvent();
-        } else if (!matchesOffset && this.#matched) {
-            this.args.action.forEach(action => action.remove(this.#target, this.#utils));
-            this.#matched = false;
-            this.callEvent();
+        if (matchesOffset !== this.#matched) {
+            this.act(matchesOffset);
+            this.#matched = matchesOffset;
+            this.callEvent(this.#matched, left, top, ...this.isOnEdge(left, top));
+            return;
+        }
+        if (this.exceededThreshold(left, top)) {
+            this.callEvent(this.#matched, left, top, ...this.isOnEdge(left, top));
+            this.#prevX = left;
+            this.#prevY = top;
         }
     }
 
-    private callEvent() {
+
+    private act(matching: boolean) {
+        if (!are(this.#action, this.#target)) {
+            return;
+        }
+        this.isLocked = true;
+        this.#action.forEach(action => {
+            if (matching) {
+                action.add(this.#target, this.#utils)
+            } else {
+                action.remove(this.#target, this.#utils)
+            }
+        });
+        this.isLocked = false;
+    }
+
+    private callEvent(matches: boolean, x: number, y: number, limitX: boolean, limitY: boolean) {
         this.emitEvent(EVENTS.OFFSET, {
             matches: this.#matched,
+            offsetX: x,
+            offsetY: y,
+            limitX: limitX,
+            limitY: limitY,
             timestamp: Date.now()
         })
     }
 
-    getRoot(): Element | Document {
-        return this.args.root ? document : this.element;
+    private getRoot(): Element {
+        return this.args.root ? document.body : this.element;
     }
+
+    private exceededThreshold(x: number, y: number): boolean {
+        return Math.abs(x - this.#prevX) > this.#threshold || Math.abs(y - this.#prevY) > this.#threshold;
+    }
+
+    private isOnEdge(x: number, y: number): [boolean, boolean] {
+        let limitY = false;
+        let limitX = false;
+        if (this.#root.scrollHeight - this.#root.clientHeight <= y + this.#threshold) {
+            limitY = true
+        }
+        if (this.#root.scrollWidth - this.#root.clientWidth <= x + this.#threshold) {
+            limitX = true
+        }
+        return [limitX, limitY];
+    }
+
+
 }