+++ src/core/listeners/move.ts
@@ -1,3 +1,4 @@
+import { TaskedEventEmitHandler } from "../bus/handlers";
 import { ICuiEventListener } from "../models/interfaces";
 import { is, round } from "../utils/functions";
 
@@ -33,6 +34,12 @@ export class CuiMoveEventListener implements ICuiEventListener<ICuiMoveEvent> {
         this.#element = element ?? document.body;
         this.#isAttached = false;
         this.#preventDefault = false;
+        this.onMouseDown = this.onMouseDown.bind(this);
+        this.onMouseUp = this.onMouseUp.bind(this)
+        this.onMouseMove = this.onMouseMove.bind(this)
+        this.onTouchStart = this.onTouchStart.bind(this)
+        this.onTouchEnd = this.onTouchEnd.bind(this)
+        this.onTouchMove = this.onTouchMove.bind(this)
     }
     setCallback(callback: (t: ICuiMoveEvent) => void): void {
         this.#onEvent = callback;
@@ -47,24 +54,26 @@ export class CuiMoveEventListener implements ICuiEventListener<ICuiMoveEvent> {
         if (this.#isAttached) {
             return;
         }
-        this.#element.addEventListener('mousedown', this.onMouseDown.bind(this))
-        this.#element.addEventListener('mouseup', this.onMouseUp.bind(this))
-        this.#element.addEventListener('mousemove', this.onMouseMove.bind(this))
-        this.#element.addEventListener('touchstart', this.onTouchStart.bind(this))
-        this.#element.addEventListener('touchend', this.onTouchEnd.bind(this))
-        this.#element.addEventListener('touchmove', this.onTouchMove.bind(this))
+        this.#element.addEventListener('mousedown', this.onMouseDown)
+        this.#element.addEventListener('mouseup', this.onMouseUp)
+        this.#element.addEventListener('mousemove', this.onMouseMove)
+        this.#element.addEventListener('touchstart', this.onTouchStart)
+        this.#element.addEventListener('touchend', this.onTouchEnd)
+        this.#element.addEventListener('touchmove', this.onTouchMove)
         this.#isAttached = true;
     }
     detach(): void {
+        console.log("On detach")
         if (!this.#isAttached) {
             return;
         }
-        this.#element.removeEventListener('mousedown', this.onMouseDown.bind(this))
-        this.#element.removeEventListener('mouseup', this.onMouseUp.bind(this))
-        this.#element.removeEventListener('mousemove', this.onMouseMove.bind(this))
-        this.#element.removeEventListener('touchstart', this.onTouchStart.bind(this))
-        this.#element.removeEventListener('touchend', this.onTouchEnd.bind(this))
-        this.#element.removeEventListener('touchmove', this.onTouchMove.bind(this))
+
+        this.#element.removeEventListener('mousedown', this.onMouseDown)
+        this.#element.removeEventListener('mouseup', this.onMouseUp)
+        this.#element.removeEventListener('mousemove', this.onMouseMove)
+        this.#element.removeEventListener('touchstart', this.onTouchStart)
+        this.#element.removeEventListener('touchend', this.onTouchEnd)
+        this.#element.removeEventListener('touchmove', this.onTouchMove)
         this.#isAttached = false;
     }